app/src/main/java/com/example/mvvmcountries/viewmodel/ListViewModel.java
package com.example.mvvmcountries.viewmodel;

import androidx.lifecycle.MutableLiveData;
import androidx.lifecycle.ViewModel;

import com.example.mvvmcountries.di.DaggerApiComponent;
import com.example.mvvmcountries.di.DaggerModuleComponent;
import com.example.mvvmcountries.model.CountriesApi;
import com.example.mvvmcountries.model.CountriesService;
import com.example.mvvmcountries.model.CountryModel;

import java.util.ArrayList;
import java.util.List;

import javax.inject.Inject;

import io.reactivex.Scheduler;
import io.reactivex.android.schedulers.AndroidSchedulers;
import io.reactivex.annotations.NonNull;
import io.reactivex.disposables.CompositeDisposable;
import io.reactivex.observers.DisposableSingleObserver;
import io.reactivex.schedulers.Schedulers;

public class ListViewModel extends ViewModel {
    public MutableLiveData<List<CountryModel>> countries = new MutableLiveData<List<CountryModel>>();
    public MutableLiveData<Boolean> countryLoadError = new MutableLiveData<Boolean>();
    public MutableLiveData<Boolean> loading = new MutableLiveData<Boolean>();

    //private CountriesService countriesService = CountriesService.getInstance();
    @Inject
    public CountriesService countriesService;

    public ListViewModel(){
        DaggerModuleComponent.create().inject(this);
    }

    private CompositeDisposable disposable = new CompositeDisposable();  //to prevent memory loss becuse service will do in background thread

    public void refresh(){
        fetchCountries();
    }


    private void fetchCountries(){
        loading.setValue(true);
        disposable.add(
                countriesService.getCountries()
                .subscribeOn(Schedulers.newThread())
                .observeOn(AndroidSchedulers.mainThread())
                .subscribeWith(new DisposableSingleObserver<List<CountryModel>>(){

                    @Override
                    public void onSuccess(@NonNull List<CountryModel> countryModels) {
                        countries.setValue(countryModels);
                        countryLoadError.setValue(false);
                        loading.setValue(false);
                    }

                    @Override
                    public void onError(@NonNull Throwable e) {
                        countryLoadError.setValue(true);
                        loading.setValue(false);
                        e.printStackTrace();
                    }
                })
        );

    }

    @Override
    protected void onCleared() {
        super.onCleared();
        disposable.clear();
    }
}

